projects/o2-timeslide-analysis/pyproject.toml
[tool.poetry]
name = "analyze_o2"
version = "0.0.1"
description = "Analyzing BBHNet outputs"
authors = ["Alec Gunny"]
license = "MIT"
# readme = "README.md"

[tool.poetry.scripts]
analyze = "analyze_o2.analysis:main"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
tqdm = "^4.64"
"bbhnet.analysis" = {path = "../../libs/analysis", develop = true}
"bbhnet.io" = {path = "../../libs/io", develop = true}
"bbhnet.logging" = {path = "../../libs/logging", develop = true}
"hermes.typeo" = "^0.1.4"

# extras required for running the notebook
jupyter = {version = "^1.0.0", optional = true}
bokeh = {version = "^2.4.2", optional = true}
rich = {version = "^11.0.0", optional = true}

[tool.poetry.extras]
notebook = ["jupyter", "bokeh", "rich"]

[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.masonry.api"
